+++ routes/api/auth.js
@@ -1,26 +1,9 @@
 const express = require("express");
 require("dotenv").config();
-
+const upload = require("../../controller/Mailer");
 const router = express.Router();
 const ctrlAuth = require("../../controller/auth");
 const ctrlContact = require("../../controller/contacts");
-const multer = require("multer");
-const path = require("path");
-const uploadDir = path.join(process.cwd(), "tmp");
-const storage = multer.diskStorage({
-	destination: (req, file, cb) => {
-		cb(null, uploadDir);
-	},
-	filename: (req, file, cb) => {
-		cb(null, file.originalname);
-	},
-	limits: {
-		fileSize: 1048576,
-	},
-});
-const upload = multer({
-	storage: storage,
-});
 
 router.post("/users/signup", ctrlAuth.signup);
 
@@ -32,7 +15,7 @@ router.get("/users/current", ctrlAuth.auth, ctrlAuth.current);
 
 router.patch(
 	"/users/avatars",
-	upload.single("picture"),
+	upload.upload.single("picture"),
 	ctrlAuth.auth,
 	ctrlContact.avatar
 );
